worker/peergrouper: eliminate JujuConnSuite usage
There was only one test in worker/peergrouper that
was using a real Mongo connection in a meaningful
way: TestInitiateReplicaSet. The other tests that
were using Mongo (via State) were doing so in
meaningless ways.

TestStartStop was taking a real State, but was not
expected to use it. There is a race though, if the
state-server machines change (not sure how?), then
the worker would be notified and would attempt to
query the mongo replicaset; in this test, mongo is
not started with --replSet, and so the test fails
in that scenario. Passing a fake State ensures we
have complete control over the state-server machines
reported to the worker.

TestPublisherSetsAPIHostPorts was using State as
a means of recording and then checking for API
host ports; it was even doing so through a mocked
publisher. Instead of passing through to State,
we report the values to the test via a channel.

TestPublisherRejectsNoServers doesn't need a State
at all, and was in the wrong place. It has been
moved and simplified.