Don't leak ErrDying out of the worker
Fixed the retry code in provisioner task to not leak out
ErrDying (by setting the error status on the machine).

In order to shutdown cleanly, the error needed to be unwrapped
as Tomb.Kill() special cases the exact object ErrDying and has
no idea about errors.Cause.  Otherwise tomb.Stop() would think
that there is a genuine error.