worker/uniter: record reboot before killing hook
When running the "juju-reboot" hook tool with --now, we
kill the hook process and then record the reboot priority.
This is racey: completion of the hook process is the
trigger for the completion of the run-hook operation.

Instead, set the reboot priority first, kill the hook
process, and the revert the reboot priority if killing
the hook process failed.