cmd/jujud/agent/machine: start using api-caller
The api-caller manifold is used with the dependency engine to maintain
an API connection. Many other manifolds will depend on it.

This change means that the machine agent is making an extra API
connection which it wasn't before. The agent will return to
maintaining just 1 API connection once the dependency engine
conversion is completed and the old "api" worker is removed.