utils: fork ReadYaml and WriteYaml from juju/utils
Juju depends on the behaviour of the yaml.v1. However this contact
is poorly expressed as the functions which perform Yaml reading and
writing are in a different repository, juju/utils, who's version of
the yaml floats.

To resolve this, fork two functions juju/utils and bring them into juju's
own utils package.

A comment explaining why this was done and the requirement on yaml.v1 has
been added to this fork.

Additionally WriteYaml has been cleaned up to avoid leaking temporary files
on disk, or overwriting a temporary file as it is being prepared.