mongo: prefer machine address for replica set peername
The network addresses used for mongo replica set initiation will
occasionally fail to resolve on MAAS resulting in a failed bootstrap.
This happens because the DNS zone tables on the MAAS controller are not
always updated with the new machine address causing subsequent lookups
of that name to fail.

If you bootstrap with --debug you see the following warning (folded for
readability):

 Initiate: fetching replication status failed:
   cannot get replica set status:
   can't get local.system.replset config from self or any seed (EMPTYCONFIG)

We workaround this MAAS issue by always preferring machine
addresses (i.e., IP addresses) for replica set initiation.

And for the MAAS provider only we drop all unresolvable hostnames from
Addresses(). If we left Addresses() with unresolvable names we end up
with:

  $ juju status
  machines:
    "0":
      agent-state: started
      agent-version: 1.26-alpha1.1
      dns-name: "maas-node3.maas"

but addressing that machine will subsequently fail:

  $ juju ssh 0
  Warning: Permanently added '10.17.17.105' (ECDSA) to the list of known hosts.
  nc: getaddrinfo: Name or service not known
  ssh_exchange_identification: Connection closed by remote host

Fixes [LP:#1412621](https://bugs.launchpad.net/juju-core/+bug/1412621)