cmd/juju/storage: fix storage commands
Some of the storage commands were panicking due to
shadowed "api" variable initialisation. Since the
API client should be scoped to the method that uses
it, don't store it on the command but store a
function that returns one. This allows mocking,
and keeps the command-execution code paths the same
for mocked and non-mocked code.