System Kill Async
This branch targets a feature branch that has not updated to the
controller name yet and still refers to system.

- On system kill, add cleanup jobs to destroy all environments and
schedule further cleanups to remove all machines and services for each
environment.

- Update state.ProcessDyingEnviron to error if the environ is a
controller with non-dead hosted environments. This ensures the
controller environ stays functional in a dying state until all hosted
environs are dead.

- Fix a bug with the undertaker worker watcher (on apiserver side) and
refactor code to make it clearer.

- Remove TestDestroyEnvironmentWithContainers as it no longer makes
sense. It's testing that we call stopInstance, which we no longer do.

- TestKillWithAPIConnection didn't actually test that the system was
killed with the api connection. The method was falling back to
destroying it via environs.Destroy, which we no longer do. So I
removed checkSystemRemovedFromStore(c, "test1", s.store).

Other notes:

I have removed destroying all services from the original
DestroyEnvironment method - as this is exactly what
cleanupServicesForDyingEnvironment does. Effectively, all
DestroyEnvironment now does is set the environment to dying and add
cleanup jobs.

assertCleanupCount has been copied to two tests, this could be moved
to a testing utils pkg (I'll see what the reviewer thinks).

Testing:

Unit tests: api, apiserver, state

Manually tested: two hosted environments and controller environment,
each with with the ubuntu service and a machine added. All machines
and services brought down successfully. Each environment can still be
switched to and status can be read, but nothing added as they are all
dead.

Follow up:

A follow up branch will poll state and display the status of the
controller take down. At the end of that, the client will call destroy
on the provider to remove the final state machines.