maas: change testing.invalid DNS name to have leading "-"
Certain ISP resolvers positively resolve "testing.invalid" but
".invalid" should not resolve at all, and it doesn't with 8.8.8.8. The
net result is that in some environments the unit tests will fail and in
others they will pass. This is fixed by changing the name to have a
leading "-" (i.e., -testing.invalid).

The tests pass on CI so they must be using an upstream resolver that
fails to resolve "testing.invalid".

Fixes [LP:#1516036](https://bugs.launchpad.net/juju-core/+bug/1516036)