Fix panic when provisioning instances on Rackspace
The rackspace provider StartInstance implementation attempts
to query InstanceConfig to validate 'firewall-mode'. This is
fine for bootstrap, but nothing ensures that setting is
propogated from Config for the state server provisioner.

This branch just switches the code the check Config directly,
and alters the testing fake so that works as expected.

An alternative would be to change instancecfg.FinishInstanceConfig
to copy the setting over, however this doesn't work neatly
as the setting is checked before delegating to the openstack
provider StartInstance which calls that function.