state: Ensure the collectionwatcher always filters by environment
collectionwatcher was recently used to implement the new
State.WatchForUnitAssignment method, but it didn't specify a filter to
limit the returned documents to a single environment. This, combined
with lax id handling inside the watcher, caused the watcher to return
"local" unit ids for docuements for the desired environment
(e.g. "ubuntu/0") and full ids for documents for other environments
(e.g. "some-long-uid:ubuntu/0"). The documents for other environments
should never have been returned.

Instead of fixing the problem in WatchForUnitAssignment,
collectionwatcher has been updated to make it impossible to repeat the
problem.

1. Environment level filtering is always applied, whether a filter
function is supplied to collectionwatcher or not.

2. If an id for the wrong environment is somehow seen by the
collectionwatcher it is now treated as an error instead of allowing it
to pass. Failing earlier and closer to the source makes the problem
easier to track down.

Side note: collectionwatcher had only previously been used for actions
and the filtering used there did limit records to a single
environment.