provider/maas: make network bridge script work for MAAS 1.9
Forward port from 1.25 to master of PR #3780.

The previous rendering of /etc/network/interfaces would result in a
broken bootstrap on MAAS 1.9 when the interface was set to DHCP. This
change reworks the script to replace interface names and bridge names
in-situ and no longer reorders the existing content; the ifup/down
scripts failed to parse the (previously) modified file.

Fixes [LP:#1512371](https://bugs.launchpad.net/juju-core/+bug/1512371)