Use cinder endpoint as given to openstack provider
Fixes lp:1512399 "environment destruction failed: ... record overflow".

Picks up fix from goose to how cinder api requests were constructed,
which was causing https requests to be sent to a http cinder endpoint.

Because the previous code also hardcoded '/v2/' in the urls being used,
logic is added in the openstack provider to prefer a v2 endpoint
where available.

As keystone does not provide a standard mechanism to specify the
version of an endpoint, this is somewhat fragile. Our Openstack
charms use a different type in the catalog for v2 api cinder, but
not all deployments will follow this scheme.