state: add watchers for remote services
Added two new methods for watching remote
services and relation unit endpoints:
 - State.WatchRemoteServices
 - Relation.WatchCounterpartEndpointUnits

The first one is necessary just so we can
watch remote services. Nothing special there.

WatchCounterpartEndpointUnits will be used
to watch the local units of a relation
involving a remote service. Because we won't
be tracking units of a remote service, we
cannot use the usual Unit -> RelationUnit ->
RelationUnit.Watch procedure. Instead, we
pass in the name of the remote service to
watch the counterpart units -- the units of
the local service involved in the relation.