apiserver/remoterelations: add facade
This is part of a new facade that will provide
access to state for an upcoming "remote relations"
worker. The worker will be responsible for
transferring settings between environments involved
in cross-model relations.

The facade is incomplete, but implements a
complex watcher that deserves isolated review.
There are two watchers: one that returns the
IDs of remote services, and one that watches
for any changes to a remote service, or any
of its counterpart relation units.

Tests have been added to featuretests that exercise
the watcher with a real State. Because the client-
side API has not been implemented yet, it is
using the facade methods directly. This will be
changed to call the client methods over the
API server when available, and then later
changed to be exercised via the worker when it
is implemented.