worker: make worker.RestartDelay a constant
Fixes LP 1519144

Many tests try to patch, or just straight overwrite the value of
worker.RestartDelay. This is unsafe because of the way the worker's
Run method operates.

This PR forces worker.RestartDelay to be a constant and allows callers
creating worker.Workers to provide the restart delay they wish.