Extract the upgradesteps worker to its own package
This change moves the upgradesteps worker from cmd/jujud/agent to
worker/upgradesteps and makes the required changes to allow this to
work. This move is required for the upcoming change to get this worker
running under the dependency engine.

A number of cleanups were made along the way including:
- a state opening function is now passed in to the worker instead of
  the worker having knowledge of how to do this.
- the function to set machine status is now passed via an interface
  instead of pulling this off the agent

The upgradesteps worker code is now cleaner but there are a number of
other improvements to make. These are documented in a new set of TODOs
at the top of worker/upgradesteps/worker.go and will be addressed in
the next PR.