new package: utils/filelock
Juju contains two types that manage on disk locks, the main one is in
juju/utils/fslock, and a second, and less featureful version in container/

This PR moves the second version to its own package (as it has nothing to do
with the container package). This move is temporary, and after refactoring
utils/filelock will be merged into juju/utils/fslock (or fslock merged into this
package).