storage: revise storage API
The storage API has been updated to remove
the legacy fields in preparation for the
upcoming 2.0 release, which affords us
an opportunity to tidy up without requiring
backwards compatibility.

In addition, various API methods have been
renamed and their arguments and results
restructured to be more consistent with
the rest of the API, and with the storage
model.