Added UUID to GCE disks
GCE Disks don't have a way to be tagged or support any other API
exported metadata.
To support UUID tagging we are using the Description field, since the
API does not support modification of said field we don't risk user
breaking it.
This could change if the metadata API grows a library or if disks start
supporting tags or other metadata form.