Don't deprecate hyphenated commands in 2.0
The initial work for hierarchical commands introduced
aliases for the old commands that were to be deprecated
in 2.0, and obsolete in 3.0.  However, the new direction
is that we will have a flat command space.

The work to flatten the commands is ongoing, but in the
mean time, the aliases for the hyphenated commands should
not be deprecated in 2.0.