state: pre-validate "--to" assignments
With the recent changes to assign units to
machines asynchronously, we now also validate
the parameters asynchronously, making for a
poor user experience. If we can determine
that an assignment cannot succeed, we should
reject the assignment immediately.

We change the state package to validate the
assignment parameters against machines specified
with "--to".

Fixes https://bugs.launchpad.net/juju-core/+bug/1524527