worker/instancepoller: Allow aggregator shutdown when request is being handled
A recent change introduced the ability of requests to the aggregator
to be aborted when the tomb is killed. This mean that sometimes a
request is sent but the reply channel is never read.

The code handling the other end of the channel wasn't changed in the
same way. There was still a naked channel write to the reply channel
meaning that the aggregator could sometimes fail to shut down. This
revision adds removed the naked write, ensuring that the tomb's Dying
channel is also checked there.