maas: add a bridge to all active interfaces
Reworked the bridge script so that all active interfaces get bridged. An
'active' interface is one that has both an 'auto' stanza and whose
'iface' method is either 'dhcp' or 'static'.

To avoid any clashes a prefix ('juju-br-') is added to device names that
become bridges; this allows for 'juju-br-eth0' and 'juju-br-bond0'.
Sticking to the old format we would end up with 'juju-br0' for both
'eth0' and 'bond0'.

I'm landing this change so that the script itself can have some exposure
on the maas-spaces branch. A subsequent PR will follow that ensures LXC
configurations are updated to include network entries for each active
interface.

The script is aware of bonds, VLANs and aliases.