maas: add bridge script using AddBootTextFile()
Add the python-based bridge script to the node using AddBootTextFile().

This is written as /tmp/add-juju-bridge.py. The run commands that use
this script are bracketed by calls that check for its existence. Ala:

  trap 'rm -f /tmp/add-juju-bridge.py' EXIT

  if [ -f /tmp/add-juju-bridge.py ]; then
     /tmp/add-juju-bridge.py ...
  fi

And the script itself is deleted once it has run once as it is not
idempotent; the script is expected to run once against the
machine-generated /etc/network/interfaces from curtin and MAAS.

This change no longer requires that the python script be encapsulated as
a bash 'here doc' which removes issues with quoting and various other
characters such as embedded "\n". And because it is no longer a 'here
doc' the script is no longer echoed to the console when cloud-init runs
during node deployment.