Make Openstack MultiSameNameEnvs compatible.
Openstack was using the environment name for certain operations which
made impossible to deploy multimple environments with the same name
with the same credentials.
Now the remaining items:
 * Instances
 * Security Groups

Have been fixed to use environment UUID instead.
Upgrade steps to change the names of the aforementioned items have been
also added.