cmd/jujud/agent: Integrate apiworkers manifold
The machine agent's workers that rely on the primary API connection
are now started using the via the apiworkers manifold instead of being
started via the machine agent's runner/worker heirarchy. This means
that the machine agent now only opens up one API connection in this
feature branch instead of having one opened by the "api-caller"
manifold and one by the (now removed) "api" worker.

This change has required significant changes in some of the machine
agent's tests due to changed semantics and the loss of the
"reportOpenedAPI" helper. Some tests were also removed because they
were testing functionality that is now part of the dependency engine
(and is well tested there).