cmd/jujud/agent: Safer channel closing in tests
Under the dependency engine workers may be restarted more often. This
causes some of the patched in functions in the machine agent's tests
to be called more than once, causing signaling-by-channel-closing to
fail after the first call (channel already closed).

This change introduces a simple "signal" struct which provides a safer
abstraction to support the test requirement of needing to know if
something was called which.