provider/maas: DNS entries do not demarcate stanzas
When parsing an interfaces file, entries beginning with 'dns-' do not
demarcate stanzas. DNS entries are now considered part of the iface
stanza that is being parsed.

For example, the following interfaces file:

   auto eth0
   iface eth0 inet dhcp

   dns-nameservers 10.10.19.2
   dns-search maas-19

we would previously bridge and render as:

   iface eth0 inet manual

   auto br-eth0
   iface br-eth0 inet dhcp
       bridge_ports eth0

   dns-nameservers 10.10.19.2
   dns-search maas-19

but is now parsed and rendered as:

   iface eth0 inet manual

   auto br-eth0
   iface br-eth0 inet dhcp
       dns-nameservers 10.10.19.2
       dns-search maas-19
       bridge_ports eth0

Fixes [LP:#1532176](https://bugs.launchpad.net/juju-core/+bug/1532176)