Store lists of image metadata atomically
The provisioner will fetch image metadata from state,
and if there is some, use that. Only if there are no
metadata matching the search criteria will it fall
back to searching the public image metadata.

Therefore, the provisioner should not be allowed to
see partial sets of image metadata, or it may not
see the subset that it is interested in. We are
currently saving image metadata one at a time; we
change it to save sets of metadata atomically.

Fixes https://bugs.launchpad.net/juju-core/+bug/1528932