cloudconfig: don't limit tools download retries
With the new Azure, instances will sometimes fail to download
tools/agent tarballs, and will thus never become active; the
machines just sit in the "idle" status forever. In Azure, it
has been observed that the amount of time that we allow is
not sufficient.

We currently limit the number of download attempts to 5, but
there's really no reason to limit the number of attempts at all.
For bootstrap, we will kill the process after a (configurable)
amount of time has elapsed. For non-bootstrap instance
creation, as mentioned above, the machine will currently just
sit idle if the 5 downloads all fail.

This change is to remove the retry limit for both Linux and
Windows, both of which have been tested live.

Fixes https://bugs.launchpad.net/juju-core/+bug/1533275