cloud: restructure cloud credentials types
It does not scale well to have concrete types
for each provider's auth-type, so Credential
has been made into a struct which contains
the raw attributes, and the auth-type constant.

Each provider will later be extended with
methods for returning the credential schemas
they support, and for automatically detecting
credentials from the environment.

An initial implementation of list-credentials
has been added to demonstrate the use of the
ParseCredentials function.

Parsing credentials no longer validates the
attributes. This enables us to lazily validate
only the credentials selected during bootstrap.