Add/implement new ProviderCredentials interface
We add environs.ProviderCredentials, an interface
containing a method whereby a provider can describe
the credentials it can accept; and a method whereby
a provider can report credentials that it is able
to automatically detect from environment variables
and files (for example).

All providers are updated to implement this interface.