state: Avoid Refresh in ModelMigration.SetPhase
As per Will's comments, ModelMigration instances should never refresh
themselves. Only the values that are actually changed by the SetPhase
are now updated on the ModelMigration's doc.

This results in a simpler, albeit sometimes less convenient, SetPhase
implementation. The required tests are also simpler.

In reality this is mostly moot as it's very likely that only one
goroutine in one process will be updating a pariticular model
migration document.