provider/maas: make dns options unique when rendering /e/n/i
When the bridge script renders an original /etc/network/interfaces file
like:

  auto eth0
      iface eth0 inet static
      gateway 10.245.168.1
      address 10.245.168.11/21
      mtu 1500

  auto eth1
  iface eth1 inet static
      dns-nameservers 10.245.168.2
      gateway 10.245.168.1
      address 10.245.168.12/21
      mtu 1500

    dns-nameservers 192.168.1.1

into:

  iface eth0 inet manual

  auto br-eth0
  iface br-eth0 inet static
      gateway 10.245.168.1
      address 10.245.168.11/21
      mtu 1500
      bridge_ports eth0

  iface eth1 inet manual

  auto br-eth1
  iface br-eth1 inet static
      dns-nameservers 10.245.168.2
      gateway 10.245.168.1
      address 10.245.168.12/21
      mtu 1500
      dns-nameservers 192.168.1.1
      bridge_ports eth1

the script needs to ensure that the dns-* options are de-duplicated and
become a single line. This knowledge gained from our understanding and
resolution to:

  https://bugs.launchpad.net/juju-core/+bug/1534795

The resolvconf(8) man page also spells this out:

To add nameserver addresses, add a single line beginning with
dns-nameservers.

  dns-nameservers 192.168.1.254 8.8.8.8

Note that one or more addresses can be given on this line, separated by
spaces, and note the "s" at the end of the option name, in contrast with
the resolv.conf(5) option name nameserver. Remember that in
interfaces(5) each option can be used at most once per stanza.

Taking the 'iface br-eth1 inet static' example it should become:

  auto br-eth1
  iface br-eth1 inet static
      gateway 10.245.168.1
      address 10.245.168.12/21
      mtu 1500
      bridge_ports eth1
      dns-nameservers 10.245.168.2 192.168.1.1

This change is also carried through to the dns-search and dns-sortlist
options.

Fixes [LP:#1536728](https://bugs.launchpad.net/juju-core/+bug/1536728)