Update bootstrap and env opening APIs
We will be getting rid of environments.yaml,
and introducing a new method of bootstrapping
environments based on a fixed "clouds.yaml",
a user-provided "credentials.yaml", and later
a (TBD) way of providing configuration to the
controller.

In preparation for this, we drop the following
functions from the environs package:
 - NewFromName
 - PrepareFromName
 - NewFromAttrs

In addition, EnvironProvider.PrepareForBootstrap
has been modified to take a new parameter struct,
which includes the existing Config value, and the
new Credential, CloudRegion, and CloudEndpoint
fields. The new fields are not yet populated or
referenced.

We also modify the metadata plugin commands to
stop writing to disk after "preparing" configuration.
These commands will need to change to accomodate
the new clouds.yaml, but for now we just want them
to not conflict with bootstrapping.