add-juju-bridge: add argument that specifies which iface to bridge
Add script arguments (--interface-to-bridge and --bridge-name) to
explicitly specify which interface should be bridged and what name it
should take. These arguments are for juju versions < 2.0 only; it allows
us to maintain compatibility for older juju versions by continuing to
use '--bridge-name=juju-br0' for the interface that is the default
route.

Examples include:

 --bridge-name=juju-br0 --interface-to-bridge=bond0
 --bridge-name=juju-br0 --interface-to-bridge=eth1