lp1536269 - Use TLS 1.0 as minimum version
For builds using < go 1.5, the default minimum
TLS version is SSLv3.  This patch will explicitly
set the minimum version to be TLS1.0

I verified that I could recreate the vulnerability
scan failure with an --upload-tools bootstrap
without this patch, then verified that the scan
did not show the vulnerabilty with this patch.