lp1538303: Retry with EOF error from API Open
During bootstrap, we may get EOF or upgrade in
progress errors while juju is still coming up.
We would retry when we got those errors from
the blocks.List call, but it is also possible
to get an EOF error from the Login (apiOpen).

Those should be retried as well.