bootstrap: providers detect regions
If you specify the name of a provider, rather than
a cloud, during bootstrap, then we will check to
see if that provider is capable of detecting regions.
If the specified provider is capable of detecting
regions, we'll ask it to do so and then use the
result to synthesise a "cloud".

In addition, we will now automatically select a
region if it is the only one, and there is not
one specified and no default either. This means
that if a provider detects a single region, then
we will use it without having to specify it by
name. We may later want to revise the detection
interface so that providers can specify which one
is the default (e.g. for lxd, that would be
"localhost", even if there are registered remotes).