apiserver: remove incorrect non error handling
logsink called primeLogSink, but then discarded the error, which it
said was not an error, but logged it at error level anyway.

There was also a test for this code, which did not exercise the error
condition - and, in working on an unrelated bug I spotted this line

    [LOG] 0:00.175 ERROR juju.apiserver Unable to prime /tmp/check-6129484611666145821/9/logsink.log (proceeding anyway): chown /tmp/check-6129484611666145821/9/logsink.log: operation not permitted

This code failed, logged, and continued on unperturbed.

Bottom line, a failure to prime the log file is either an error, and
the code should return at this point -- or -- priming the log file is
not necessary, so we can just delete the whole thing.