cmd/juju/commands/bootstrap: always test sentinel value when retrying
Update LP#1538583

During bootstrap cmd/juju would inspect the string value of an error to
determine if the error was one it could retry or not. All the errors it
was testing against have known sentinel values, so use those instead.

Also, unify all the places that apiserver.UpgradeInProcessError is used
rather than strings which look the same, plus some other information for
the user. If the text is to be changed, now we'll change it only in one
place.