lp1541473 - Resize certChangedChan
It is possible for the certupdater to come
up before the apiserver does, and to send
multiple updates for the certificate.  The
certChangedChan should have a larger size
to accommodate these requests.

I chose 10 as we can allow up to 7 state
servers, and we may update the cert for
each server, plus another update to account
for addresses local to machine 0 (::1,
127.0.0.1, etc).