Move suite_test.go to the peergrouper package
Tests were not running in this directory because
suite_test.go was the only file in the
peergrouper_test package.  Updated that file, and
added in mock information for the new watcher
and machine status methods added in the fix for
bug 1541473.  Unit tests need to be added and that
work is tracked in bug 1543408.