cloudconfig/sshinit: don't send commands to stdin
Turns out that bash reads byte-at-a-time when reading
commands from stdin. This is likely so it can avoid
buffering more than a command's worth of input.

Anyway, we send huge scripts to bash at bootstrap
time, particularly when uploading tools. With this
change, we avoid sending the script to bash directly,
and instead write it to the remote disk and execute
it from there.

Fixes https://bugs.launchpad.net/juju-core/+bug/1543388