cloud: add StorageEndpoint to Cloud/Region
Add storage-endpoint to the cloud definition, which
is needed by Azure. Azure has separate endpoints for
storage and everything else, and you can't derive
one from the other.

Also updated the azure cloud definition in
fallback-public-cloud.yaml, and updated the Azure
provider to use the new field. In the YAML file
the endpoint is specified as a URL for consistency,
despite the Azure code only needing the host name
to use as a domain-name suffix.