cmd/juju/commands: allow bootstrap with no regions
If a cloud defines no regions, and no region is
specified, we should just pass an empty region
name to the provider, with the cloud's endpoint
values.

This enables the definition of a MAAS cloud with
no regions. i.e. with this change you can have:

    clouds:
      my-maas:
        type: maas
        auth-types: [oauth1]
        endpoint http://maas-host/MAAS

Also reinstate maas-server and maas-oauth config
attributes. They need to be there or we get spurious
warnings about unknown attributes.