provider/manual: take host from region/endpoint
The manual provider is updated to support DetectRegions,
which just returns "NotFound" to indicate that there are
no pre-defined regions. This allows the provider to be
used without a clouds.yaml entry.

In addition, we now support taking the bootstrap-host
value from the endpoint (if specified in clouds.yaml),
or the "region name" (if specified on the command line
with no corresponding clouds.yaml entry).

i.e. you can now do:
    juju bootstrap manual/<host>