Write jujuclient files on bootstrap
In evirons.Prepare and environs.Destroy, we now
update the jujuclient store accordingly. Because
most code is still using the configstore code,
we have to revert to giving the initial model
the same name as the controller for now.

Some minor fallout:
 - UserManager API client code now accepts
   non-local user tags, and leaves it to the
   server to reject non-local ones. This
   allows one to pass "user@local" to those
   client methods, as well as just "user".
 - Removed some test code that allowed the
   admin user to be something other than
   "admin@local". This was only used by the
   dummy provider, which called the user
   "dummy-admin@local". Why is a mystery.
 - Remove some dead code in destroy-controller
   and kill-controller.
 - Drop environs.DestroyInfo, and update the
   one external caller, destroy-model to just
   destroy the EnvironInfo and also remove the
   model from the jujuclient store.

Also updated the change-user-password command
to operate off jujuclient, but still update
configstore for now. While I was at it, I
removed the .server file code.