lp1544796: Fix error checking in Restore
Bug 1544796 was introduced with PR
https://github.com/juju/juju/pull/4361, which
changed how errors are returned from the rpc
calls.  This spot where the error was compared
to rpc.ErrShutdown was not updated to use
errors.Cause() as in other places.

There were some other logic errors in the restore
code which needed to be addressed:

1 - The check for an "upgrade in progress" was wrong.
That error comes from the rpc call, and not as the
result from the method on the facade being called.  So,
the check should be if err is an "upgrade in progress"
error, not remoteError.  We also cannot use
params.IsCodeUpgradeInProgress as that will not work
on the rpc.RequestError error type that is returned.
The check now uses the same method the bootstrap command
does.

2 - In restore() and finishRestore(), err was
re-declared inside the attempt strategy block, but then
used outside of the block where it would always have
a nil value.