Improved "switch" command
The switch command has been updated to use the new
jujuclient infrastructure, and to automatically
refresh the local model cache if the requested
model is not known locally.

This means you can do (once add-user is updated):
    (as admin) juju add-user bob --share admin
    (as bob)   juju register <registration-string>
               juju switch admin

As part of this work, the cleanup of the old ways
has continued. Notably, we have removed all code
related to the "current-model" file. The current
model is now stored only in the jujuclient store,
and the current controller is recorded in the
current-controller file. The primary difference
is that previously the current-controller file
would be removed when switching to a model, and
now we will not.

Note that the switch command's ability to show
the current controller/model, and list the cache,
have been removed. Listing is now handled by the
list-controllers and list-models commands. We
have yet to implement a replacement for showing
the current, but it is likely that we will add
current-model and current-controller commands.